step1수정: 파일분리 및 person init 수정
